requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
Automat==24.8.1
autopep8==2.3.2
cffi==1.17.1
constantly==23.10.4
cryptography==44.0.1
css-html-js-minify==2.5.5
Django==5.1.7
django-cors-headers==4.7.0
django-crispy-forms==2.0
django-extensions==3.2.3
django-formtools==2.5.1
django-otp==1.6.0
django-phonenumber-field==8.0.0
django-two-factor-auth==1.17.0
flake8==7.2.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
mccabe==0.7.0
msgpack==1.1.0
mysqlclient==2.1.0
packaging==24.2
phonenumbers==9.0.2
pillow==11.2.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.13.0
pycparser==2.22
pyflakes==3.3.2
pyOpenSSL==25.0.0
pyotp==2.9.0
pypng==0.20220715.0
pytz==2025.1
qrcode==7.4.2
redis==5.2.1
service-identity==24.2.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
whitenoise==6.3.0
zope.interface==7.2
